ml_requirements.txt
pandas
numpy
scikit-learn
xgboost
requests
sqlalchemy
psycopg2-binary
openmeteo-requests
python-dotenv
geopandas
shapely
matplotlib
seaborn
hdbscan
